+++ lightning/src/ln/mod.rs
@@ -16,6 +16,7 @@ pub mod router;
 pub mod peer_handler;
 pub mod chan_utils;
 pub mod features;
+pub(crate) mod onchaintx;
 
 #[cfg(feature = "fuzztarget")]
 pub mod peer_channel_encryptor;
